// web/projects/portal/src/app/vsobjects/model/schedule/simple-schedule-dialog-model.ts
/*
 * inetsoft-web - StyleBI is a business intelligence web application.
 * Copyright © 2024 InetSoft Technology (info@inetsoft.com)
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
import {
   TimeConditionModel,
   TimeRange
} from "../../../../../../shared/schedule/model/time-condition-model";
import { TimeZoneModel } from "../../../../../../shared/schedule/model/time-zone-model";
import { CommonKVModel } from "../../../common/data/common-kv-model";
import { EmailAddrDialogModel } from "../../../widget/email-dialog/email-addr-dialog-model";
import { ActionModel } from "../../../widget/schedule/action-model";
import { CSVConfigModel } from "../../../../../../shared/schedule/model/csv-config-model";

export interface SimpleScheduleDialogModel {
   userDialogEnabled: boolean;
   timeProp: string;
   twelveHourSystem: boolean;
   taskName: string;
   isSecurity: boolean;
   formatTypes: CommonKVModel<string, string>[];
   expandEnabled: boolean;
   emailButtonVisible: boolean;
   emailDeliveryEnabled: boolean;
   timeConditionModel: TimeConditionModel;
   actionModel: ActionModel;
   emailAddrDialogModel: EmailAddrDialogModel;
   timeRanges: TimeRange[];
   timeZoneOptions?: TimeZoneModel[];
   startTimeEnabled: boolean;
   timeRangeEnabled: boolean;
   users?: string[];
   groups?: string[];
   emailGroups?: string[];
   tableAssemblies?: string[];
}